Line breaks expressed as {.n}
A space is left on either side of the curly braces when the line break corresponds to a word boundary (most cases). The break that begins line 7 occurs in the middle of a word so spaces are used around the curly braces (the hyphen in the original print version is not kept since it just also indicated a line break).

Also testing whether "curly quotes" get by.